refactor: Rename folder 'db' to 'database' for improved code semantics
