+++ core/database/Initialize.go
@@ -1,9 +1,9 @@
-package db
+package database
 
 import (
 	"fmt"
 
-	"github.com/jyotirmoydotdev/openfy/db/models"
+	"github.com/jyotirmoydotdev/openfy/database/models"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 )
@@ -11,11 +11,11 @@ import (
 var dbInstance *gorm.DB
 var productdbInstance *gorm.DB
 
-func GetDB() (*gorm.DB, error) {
+func GetCustomerDB() (*gorm.DB, error) {
 	if dbInstance != nil {
 		return dbInstance, nil
 	}
-	db, err := gorm.Open(sqlite.Open("./db/databaseCustomerStaffMember.db"), &gorm.Config{})
+	db, err := gorm.Open(sqlite.Open("./database/databaseCustomerStaffMember.db"), &gorm.Config{})
 	if err != nil {
 		return nil, fmt.Errorf("error opening database: %v", err)
 	}
@@ -27,7 +27,7 @@ func GetProductDB() (*gorm.DB, error) {
 	if productdbInstance != nil {
 		return productdbInstance, nil
 	}
-	db, err := gorm.Open(sqlite.Open("./db/databaseProduct.db"), &gorm.Config{})
+	db, err := gorm.Open(sqlite.Open("./database/databaseProduct.db"), &gorm.Config{})
 	if err != nil {
 		return nil, fmt.Errorf("error opening database: %v", err)
 	}
@@ -36,7 +36,7 @@ func GetProductDB() (*gorm.DB, error) {
 }
 
 func InitializeDatabases() error {
-	db, err := GetDB()
+	db, err := GetCustomerDB()
 	if err != nil {
 		return err
 	}
