+++ core/database/models/productModel.go
